Bug 1838983: When dispatching a focus event, clear the stored active item if it is different to the target. r=eeejay
When the Accessible is recreated for the focused node, DocAccessible::CreateSubtree calls FocusManager::DispatchFocusEvent directly.
If FocusManager::mActiveItem was set, DispatchFocusEvent previously didn't clear it, even though DispatchFocusEvent was given a new target overriding mActiveItem.
This meant that the old mActiveItem would remain until it was next set or cleared, causing assertions if it died and potentially other problems.
To fix this, DispatchFocusEvent clears mActiveItem if it is different to the target, since the target should override.

Differential Revision: https://phabricator.services.mozilla.com/D181605

UltraBlame original commit: d0376dca2d43daf1c2ff860b385586953d36e332