Bug 1838428 - Use Python http module for benchmark server. r=perftest-reviewers,kshampur
This patch replaces the wptserve HTTP server with the builtin python http module. The reason for this is that the wptserve module can't be easily fixed/modified to suit our needs, and the mozhttpd server is based on `six` for backwards compatibility with Python 2, but we don't need that anymore. Furthermore, mozhttpd is deprecated even though it is still in use in a few areas. In this case, the http module provides a feature that most servers provide by default which is that a request for a directory automatically finds the index file in the directory and returns those contents.

Differential Revision: https://phabricator.services.mozilla.com/D181461

UltraBlame original commit: b6233acd02848323203feb0864cfc994aaa0f4f3