Bug 831008 - Make `EventListenerManager` assert when a legacy DOM mutation event listener is added in either chrome documents, native anonymous subtrees or the system event groups r=smaug
The legacy DOM mutation events shouldn't be used even in content and must not be
used in chrome.

This patch makes it crash if legacy mutation event listener is:
* added to a node in a chrome document
* added to a node in a native anonymous subtree
* added to the system event group
* implemented by C++

This causes some crash tests and chrome tests.  Therefore, this patch modifies
them to use `MutationObserver` instead or removes them if it's impossible to
rewrite with `MutationObserver`.

Note that if I prevent to dispatch mutation events in some/all of the cases,
that causes performance regression in Speedometer3.  Therefore, this patch
does not touch `EventDispatcher`.

Differential Revision: https://phabricator.services.mozilla.com/D181828

UltraBlame original commit: 992d6e7accc80b3ce8adaa771e5a79626ffed587