Bug 1838011 - [devtools] Handle CSS Nesting in parseDeclarations. r=devtools-reviewers,ochameau.
This adds different nested state so we can simply ignore nested CSS rules.
I added a lot of different test cases in the existing `parseDeclarations`
unit test so we can have good coverage.

It also removes "{" from `CSS_BLOCKS` as this wasn't something that could happen
before nesting, and we're handling that separately.
We took this as an opportunity to put opening brackets in `currentBlocks` instead
of closing ones, as it's a bit easier to understand.

Differential Revision: https://phabricator.services.mozilla.com/D180701

UltraBlame original commit: ac83781a23205e68fd58b5b16199958b6a07308d