Bug 1837620 - Part 6: Make edges for multiple shape guard weak too r=sfink
To make the edges for the multiple shape guard weak we need a way to find and
update the objects used to hold these edges. The simplest way to do this is to
add a vector in the zone.  I don't believe that there are not huge numbers of
these objects.

Differential Revision: https://phabricator.services.mozilla.com/D182012

UltraBlame original commit: b6d68da027a23da0a86fd73efc2fc004a2e763ba