Bug 1841250 - Always skip first code page on Windows 64-bit platforms. r=iain
On Win64 platforms (`NEED_JIT_UNWIND_HANDLING`), we reserve an extra page in
`ReserveProcessExecutableMemory` for the generated exception handler.

Before this patch we'd skip the first page if we generated an exception handler there.
If we didn't generate an exception handler (for example JS shell builds on ARM64)
we'd not skip the first page and instead have an unused page at the end of the JIT
code region.

With this patch we always skip the first page if we reserved one. This fixes an
assertion failure in `UnregisterJitCodeRegion` for Windows ARM64 JS shell builds
because the size didn't match what we passed to `RegisterJitCodeRegion`.

Differential Revision: https://phabricator.services.mozilla.com/D182726

UltraBlame original commit: e810f2c019bac01c5a7e2e5807841701db9fa67d