Bug 1822310 - [devtools] Move from "context" to generatedSourceActor assertion for ADD_ORIGINAL_SOURCES. r=devtools-reviewers,bomsy
As we declare a related generated source actor, we can ensure it is still valid in the reducer.
The source actors are being removed on thread removal.

For ADD_SOURCES, it is harder to find a good assertion.
The current context assertion is useless as we fetch the "cx" just before dispatching the ADD_SOURCES action...
Anyway, this action is emitted synchronously from onResourceAvailable so that should be reliable.
The only check we could do would be to ensure that the thread which emitted these SOURCE resources
is still registered in the reducer, but that's a bit complex as wight be notified about sources
related to many threads (via ResourceCommand throttling).

Differential Revision: https://phabricator.services.mozilla.com/D182968

UltraBlame original commit: cfc433febf5b9f7ea095c4ef1f7f82b0e71d3b0e