Bug 1843210 - Allow options with choices without value. r=firefox-build-system-reviewers,sergesanspaille
I want to make --enable-elf-hack the (kind of) default, but also allow
to specify --enable-elf-hack=legacy or --enable-elf-hack=relr to give a
specific type of ELF hack to apply, and --disable-elf-hack would
explicitly disable it. That's currently not allowed, although it's kind
of possible to fake it (setting --disable-elf-hack as the default), but
then --disable-elf-hack would not appear in the help.

Differential Revision: https://phabricator.services.mozilla.com/D183452

UltraBlame original commit: e93250caba655cc43483398bf34b942f3581a05f