Bug 1796699 - Allow for a configurable alt-tag on the about:welcome/Spotlight logo element r=omc-reviewers,aminomancer
Allows a localized alt tag to be added to the logo element (for ex. in Spotlights where the logo is replaced by an image.)

Implemented similarly to the alt-tag for the onboarding illustrations; localization requires a `string_id` with an associated `.aria-label` prop.

Differential Revision: https://phabricator.services.mozilla.com/D184139

UltraBlame original commit: ffe09128b9f7ec871feaa1803545e8ebae4139c7