Bug 1844243 - Statically link libcrypto to xar. r=firefox-build-system-reviewers,sergesanspaille
When it's not built on the same kind of docker image that runs it,
there's a possibility of incompatibility because of the recent
change in openssl library version. This happens when crossing between
debian 11 and debian 12, for instance.

Differential Revision: https://phabricator.services.mozilla.com/D183944

UltraBlame original commit: b5271e1aac5d5ca657a2f1ba978d1314e47cfa9e