Bug 1843749 - Block qcom_tiled_rendering on Adreno 308 GPUs. r=gfx-reviewers,lsalzman
We started using this extension as an optimization on Adreno GPUs in
bug 1828248. At the time, we discovered that there was a driver bug in
version 0490 of the Adreno driver resulting in rendering errors on a
variety of Adreno GPUs. We therefore blocked the usage of the
extension on that driver version.

We have now discovered another bug causing even more apparent
rendering issues. This appears to only affect Adreno 308 GPUs running
driver versions 331 or 415. This patch therefore additionally blocks
the extension on such devices.

Differential Revision: https://phabricator.services.mozilla.com/D184388

UltraBlame original commit: 12931a93e28c885a13a341c5dde0858355ed8038